Moved ast-generators into separate module
A centralized ast-generator module should make the reuse of repeating structures easier and hide the actual code-generation from plugin-developers.

Optimized imports
