+++ turms/plugins/base.py
@@ -1,10 +1,12 @@
 from abc import abstractmethod
 import ast
-from typing import List, Union, Type, TypeVar, Tuple, Dict, Optional
+from typing import List, Union, Type, TypeVar, Tuple, Dict
 
 from graphql import GraphQLNamedType
 from pydantic import BaseModel, BaseSettings
 from graphql.utilities.build_client_schema import GraphQLSchema
+
+from turms.ast_generators import TypeAstGenerator
 
 GNT = TypeVar("GNT", bound=GraphQLNamedType)
 
@@ -51,72 +53,4 @@ class Plugin(BaseModel):
 
     @staticmethod
     def _generate_type_description_ast(gql_type: GraphQLNamedType) -> ast.Expr:
-        return AstGenerator.generate_type_description(gql_type.description)
-
-
-class AstGenerator:
-    @staticmethod
-    def _generate_description(description: str) -> ast.Expr:
-        return ast.Expr(value=ast.Constant(value=description))
-
-    @staticmethod
-    def generate_type_description(
-            description: str
-    ) -> ast.Expr:
-        return AstGenerator._generate_description(description)
-
-    @staticmethod
-    def generate_field_description(
-            description: str
-    ) -> ast.Expr:
-        return AstGenerator._generate_description(description)
-
-    @staticmethod
-    def generate_field_alias(original_name: str) -> ast.Call:
-        return ast.Call(
-            func=ast.Name(id="Field", ctx=ast.Load()),
-            args=[],
-            keywords=[
-                ast.keyword(arg="alias", value=ast.Constant(value=original_name))
-            ]
-        )
-
-    @staticmethod
-    def generate_attribute(
-            name: str,
-            annotation: ast.AST,
-            alias: Optional[ast.Call] = None
-    ) -> ast.AnnAssign:
-        return ast.AnnAssign(
-            target=ast.Name(name, ctx=ast.Store()),
-            annotation=annotation,
-            value=alias,
-            simple=1
-        )
-
-    @staticmethod
-    def generate_enum_value(
-            field_name: str,
-            enum_value: str
-    ) -> ast.Assign:
-        return ast.Assign(
-            targets=[ast.Name(id=field_name, ctx=ast.Store())],
-            value=ast.Constant(value=enum_value)
-        )
-
-    @staticmethod
-    def generate_class_definition(
-            classname: str,
-            base_classes: List[str],
-            body: List[ast.AST]
-    ) -> ast.ClassDef:
-        return ast.ClassDef(
-            classname,
-            bases=[
-                ast.Name(id=base.rsplit(".", maxsplit=1)[-1], ctx=ast.Load())
-                for base in base_classes
-            ],
-            decorator_list=[],
-            keywords=[],
-            body=body,
-        )
+        return TypeAstGenerator.generate_type_description(gql_type.description)
